Add utils and clean ga code
